Optimized main.py
Optimized the message receiving system. Instead of refreshing the buffer via iterating through a list of messages, it will instead initially iterate through those messages and when a new message is detected it will append the message to three 3 lists. messages_raw, messages and messages_str. Each hold their own purpose in the code, and this allows me to further optimize the code by not relying on built-in Python for loops but rather the .join function.